Refactor backend code into directories
