+++ backend/Todos.py
@@ -2,11 +2,11 @@ from typing import List, Tuple
 
 from fastapi import HTTPException
 
-from IOSchema import TodoInput, TodoUpdate, TodoDetails
-from OrganisationHelpers import getOrganisationByName
+from backend.States.IOSchema import TodoInput, TodoUpdate, TodoDetails
+from backend.Organisation.OrganisationHelpers import getOrganisationByName
 from UserAccounts import getUserByID
 from Meeting import Task
-from database import updateTodos, addTodos, getMeetingTodos, getUserTodosOrg, getUserOrgs, getUserTodos, \
+from backend.database.database import updateTodos, addTodos, getMeetingTodos, getUserTodosOrg, getUserOrgs, getUserTodos, \
     replaceMeetTodos
